auth: some refactoring
