+++ src/backend/src/auth/modules/jwt/interfaces/jwt-payload.interface.ts
@@ -1,10 +1,4 @@
-import { OAuthUserDto } from '../../oauth/dto';
-
 export interface JwtPayload {
   sub: number; // user id
   username: string;
-}
-
-export interface OAuthJwtPayload {
-  oAuthUser: OAuthUserDto;
 }
