+++ src/backend/src/auth/modules/oauth/oauth.controller.ts
@@ -1,6 +1,5 @@
 import { Body, Controller, Get, Post, UseGuards } from '@nestjs/common';
 import { User } from '../../decorators';
-import { JwtAuthService } from '../jwt/jwt-auth.service';
 import { OAuthUserDto, OAuthSignUpDto } from './dto';
 import { OAuth2Guard } from './guard';
 import { OauthService } from './oauth.service';
@@ -9,7 +8,7 @@ import { OauthService } from './oauth.service';
 export class OauthController {
   private readonly client_id = process.env['OAUTH_42_CLIENT_ID'];
 
-  constructor(private service: OauthService, private jwt: JwtAuthService) {}
+  constructor(private service: OauthService) {}
 
   @Get('client_id')
   getOAuthClientId() {
@@ -31,7 +30,7 @@ export class OauthController {
   @UseGuards(OAuth2Guard)
   async oauthSignUpTempToken(@User() user: OAuthUserDto) {
     return {
-      jwt: await this.jwt.signTempToken({ oAuthUser: user }),
+      jwt: await this.service.signTempToken({ oAuthUser: user }),
     };
   }
 }
