+++ src/backend/src/auth/modules/oauth/oauth.module.ts
@@ -4,12 +4,12 @@ import { OauthController } from './oauth.controller';
 import { AuthUtilsModule } from '../utils/auth-utils.module';
 import { DbModule } from 'src/db/db.module';
 import { HttpModule } from '@nestjs/axios';
-import { JwtAuthModule } from '../jwt/jwt-auth.module';
 import { OAuth2Strategy } from './strategy';
+import { JwtModule } from '@nestjs/jwt';
 
 @Module({
   providers: [OauthService, OAuth2Strategy],
   controllers: [OauthController],
-  imports: [AuthUtilsModule, DbModule, HttpModule, JwtAuthModule],
+  imports: [AuthUtilsModule, DbModule, HttpModule, JwtModule],
 })
 export class OauthModule {}
