--- src/frontend/src/app/auth/interface/auth.interface.ts
@@ -1,9 +1,0 @@
-export interface OAuthUser {
-  login: string;
-  email: string;
-  image_url: string;
-}
-
-export interface OAuthJwtPayload {
-  oAuthUser: OAuthUser;
-}
