--- src/frontend/src/app/auth/interface/index.ts
@@ -1,1 +1,0 @@
-export * from './auth.interface';
