+++ src/frontend/src/app/auth/oauth/dto/auth.dto.ts
@@ -1,0 +1,15 @@
+export interface OAuthUserDto {
+  login: string;
+  email: string;
+  image_url: string;
+}
+
+// TODO: use it in signup.service.ts
+// export interface OAuthSignUpDto {
+//   username: string;
+//   jwt: string;
+// }
+
+export interface OAuthJwtPayload {
+  oAuthUser: OAuthUserDto;
+}
