+++ src/frontend/src/app/auth/oauth/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './auth.dto';
